clean out old code to avoid confusing repeated functions
