+++ bigraph_viz/__init__.py
@@ -1,7 +1,6 @@
 import pprint
-from bigraph_viz.diagram import plot_bigraph, generate_types
+from bigraph_viz.visualize_types import VisualizeTypes
 from bigraph_viz.dict_utils import replace_regex_recursive
-from bigraph_schema import TypeSystem
 
 
 pretty = pprint.PrettyPrinter(indent=2)
