+++ bigraph_viz/visualize_types.py
@@ -1,96 +1,31 @@
-"""
-Bigraph diagram
-"""
 import os
-from bigraph_schema import TypeSystem, Edge
-from bigraph_viz.plot_old import absolute_path, make_label, check_if_path_in_removed_nodes
 import graphviz
 
+from bigraph_schema import TypeSystem, is_schema_key
+from bigraph_viz.dict_utils import absolute_path
+
+
+REMOVE_KEYS = ['global_time']
 
 PROCESS_SCHEMA_KEYS = [
     'config',
     'address',
     'interval',
     'inputs',
-    'outputs', 
+    'outputs',
     'instance',
 ]
 
-REMOVE_KEYS = ['global_time']
 
-updated_path_type = {
-        '_type': 'path',
-        '_inherit': 'list[string]~string',
-        # '_apply': apply_path
-}
-
-step_type = {
-    '_type': 'step',
-    '_inherit': 'edge',
-    'address': 'string',
-    # '_default': {
-    #     'inputs': {},
-    #     'outputs': {}},
-    # '_inputs': 'string~tuple',
-    # '_outputs': 'string~tuple',
-    'config': 'schema'}
+def make_label(label):
+    # label = label.replace(' ', '<br/>')  # replace spaces with new lines
+    return f'<{label}>'
 
 
-process_type = {
-    '_type': 'process',
-    '_inherit': 'step',
-    'interval': 'float'}
-
-
-composite_type = {
-    '_type': 'composite',
-    '_inherit': 'process',
-    'bridge': 'wires',
-    # 'config': 'schema'
-}
-
-def generate_types():
-    core = TypeSystem()
-    core.register('path', updated_path_type, strict=False)
-    core.register('step', step_type, strict=False)
-    core.register('process', process_type, strict=False)
-    core.register('composite', composite_type, strict=False)
-    return core
-
-
-def plot_edges(
-        graph,
-        edge,
-        port_labels,
-        port_label_size,
-        state_node_spec,
-        constraint='false',
-):
-    process_path = edge['edge_path']
-    process_name = str(process_path)
-    target_path = edge['target_path']
-    port = edge['port']
-    target_name = str(target_path)
-
-    # place it in the graph
-    # TODO -- not sure this is working, it might be remaking the node
-    if target_name not in graph.body:  # is the source node already in the graph?
-        label = make_label(target_path[-1])
-        graph.node(target_name, label=label, **state_node_spec)
-
-    # port label
-    label = ''
-    if port_labels:
-        label = make_label(port)
-
-    with graph.subgraph(name=process_name) as c:
-        c.edge(
-            target_name,
-            process_name,
-            constraint=constraint,
-            label=label,
-            labelloc="t",
-            fontsize=port_label_size)
+def check_if_path_in_removed_nodes(path, remove_nodes):
+    if remove_nodes:
+        return any(remove_path == path[:len(remove_path)] for remove_path in remove_nodes)
+    return False
 
 
 def get_graph_wires(
@@ -156,6 +91,36 @@ def get_graph_wires(
 
     return graph_dict
 
+def plot_edges(
+        graph,
+        edge,
+        port_labels,
+        port_label_size,
+        state_node_spec,
+        constraint='false',
+):
+    process_path = edge['edge_path']
+    process_name = str(process_path)
+    target_path = edge['target_path']
+    port = edge['port']
+    target_name = str(target_path)
+    # place it in the graph
+    # TODO -- not sure this is working, it might be remaking the node
+    if target_name not in graph.body:  # is the source node already in the graph?
+        label = make_label(target_path[-1])
+        graph.node(target_name, label=label, **state_node_spec)
+    # port label
+    label = ''
+    if port_labels:
+        label = make_label(port)
+    with graph.subgraph(name=process_name) as c:
+        c.edge(
+            target_name,
+            process_name,
+            constraint=constraint,
+            label=label,
+            labelloc="t",
+            fontsize=port_label_size)
 
 def get_graph_dict(
         schema,
@@ -295,7 +260,6 @@ def get_graphviz_fig(
     node_groups = node_groups or []
     node_names = []
     invisible_edges = invisible_edges or []
-
     # node specs
     state_node_spec = {
         'shape': 'circle', 'penwidth': '2', 'margin': label_margin, 'fontsize': node_label_size}
@@ -307,35 +271,32 @@ def get_graphviz_fig(
         'style': 'dashed', 'penwidth': '1', 'arrowhead': 'normal', 'arrowsize': '1.0', 'dir': 'forward'}
     output_edge_spec = {
         'style': 'dashed', 'penwidth': '1', 'arrowhead': 'normal', 'arrowsize': '1.0', 'dir': 'back'}
-
     # initialize graph
     graph = graphviz.Digraph(name='bigraph', engine='dot')
     graph.attr(size=size, overlap='false', rankdir=rankdir, dpi=dpi)
-
     # state nodes
     graph.attr('node', **state_node_spec)
     for node in graph_dict['state_nodes']:
         node_path = node['path']
         node_name = str(node_path)
         node_names.append(node_name)
-
         # make the label
         label = node_path[-1]
-        schema_label = ''
+        schema_label = None
         if show_values:
-            if node.get('value') is not None:
-                schema_label += f"<br/> {node['value']}"
+            if node.get('value'):
+                if not schema_label:
+                    schema_label = ''
+                schema_label += f": {node['value']}"
         if show_types:
             if node.get('type'):
-                if schema_label == '':
-                    schema_label += f"<br/> [{node['type']}]"
-                else:
-                    schema_label += f" : [{node['type']}]"
+                if not schema_label:
+                    schema_label = '<br/>'
+                schema_label += f"[{node['type']}]"
         if schema_label:
             label += schema_label
         label = make_label(label)
         graph.node(str(node_name), label=label)
-
     # process nodes
     process_paths = []
     graph.attr('node', **process_node_spec)
@@ -346,29 +307,24 @@ def get_graphviz_fig(
         node_names.append(node_name)
         label = make_label(node_path[-1])
         graph.node(node_name, label=label)
-
     # place edges
     graph.attr('edge', arrowhead='none', penwidth='2')
     for edge in graph_dict['place_edges']:
-
         # show edge or not
         show_edge = True
         if remove_process_place_edges and edge['child'] in process_paths:
             show_edge = False
         elif edge in invisible_edges:
             show_edge = False
-
         if show_edge:
             graph.attr('edge', style='filled')
         else:
             graph.attr('edge', style='invis')
-
         parent_node = str(edge['parent'])
         child_node = str(edge['child'])
         graph.edge(parent_node, child_node,
                    dir='forward', constraint='true'
                    )
-
     # input edges
     for edge in graph_dict['input_edges']:
         if edge['type'] == 'bridge_inputs':
@@ -377,7 +333,6 @@ def get_graphviz_fig(
         else:
             graph.attr('edge', **input_edge_spec)
             plot_edges(graph, edge, port_labels, port_label_size, state_node_spec, constraint='true')
-
     # output edges
     for edge in graph_dict['output_edges']:
         if edge['type'] == 'bridge_outputs':
@@ -386,33 +341,26 @@ def get_graphviz_fig(
         else:
             graph.attr('edge', **output_edge_spec)
             plot_edges(graph, edge, port_labels, port_label_size, state_node_spec, constraint='true')
-
     # disconnected input edges
     for edge in graph_dict['disconnected_input_edges']:
         process_path = edge['edge_path']
         port = edge['port']
-
         # add invisible node for port
         node_name2 = str(absolute_path(process_path, port)) + '_input'
         graph.node(node_name2, label='', style='invis', width='0')
         edge['target_path'] = node_name2
-
         graph.attr('edge', **input_edge_spec)
         plot_edges(graph, edge, port_labels, port_label_size, state_node_spec, constraint='true')
-
     # disconnected output edges
     for edge in graph_dict['disconnected_output_edges']:
         process_path = edge['edge_path']
         port = edge['port']
-
         # add invisible node for port
         node_name2 = str(absolute_path(process_path, port)) + '_output'
         graph.node(node_name2, label='', style='invis', width='0')
         edge['target_path'] = node_name2
-
         graph.attr('edge', **output_edge_spec)
         plot_edges(graph, edge, port_labels, port_label_size, state_node_spec, constraint='true')
-
     # grouped nodes
     for group in node_groups:
         group_name = str(group)
@@ -429,7 +377,6 @@ def get_graphviz_fig(
                     previous_node = node_name
                 else:
                     print(f'node {node_name} not in graph')
-
     # formatting
     if node_border_colors:
         for node_name, color in node_border_colors.items():
@@ -437,9 +384,7 @@ def get_graphviz_fig(
     if node_fill_colors:
         for node_name, color in node_fill_colors.items():
             graph.node(str(node_name), color=color, style='filled')
-
     return graph
-
 
 def plot_bigraph(
         state,
@@ -481,12 +426,10 @@ def plot_bigraph(
     remove_nodes = kwargs.pop('remove_nodes')
     show_process_schema_keys = kwargs.pop('show_process_schema_keys')
     remaining_kwargs = dict(kwargs)
-
     # set defaults if none provided
-    core = core or generate_types()
+    core = core or VisualizeTypes()
     schema = schema or {}
     schema, state = core.generate(schema, state)
-
     # parse out the network
     graph_dict = get_graph_dict(
         schema=schema,
@@ -495,10 +438,8 @@ def plot_bigraph(
         remove_nodes=remove_nodes,
         show_process_schema_keys=show_process_schema_keys,
     )
-
     # make a figure
     graph = get_graphviz_fig(graph_dict, **remaining_kwargs)
-
     # display or save results
     if print_source:
         print(graph.source)
@@ -511,11 +452,203 @@ def plot_bigraph(
     return graph
 
 
+
+
+
+
+
+
+
+
+def graphviz_any(core, schema, state, path, options, graph):
+    if len(path) > 0:
+        node_spec = {
+            'name': path[-1],
+            'path': path,
+            'value': None,
+            'type': core.representation(schema)}
+
+        if not isinstance(state, dict):
+            node_spec['value'] = state
+
+        graph['state_nodes'].append(node_spec)
+
+    if isinstance(state, dict):
+        for key, value in state.items():
+            if not is_schema_key(key):
+                subpath = path + (key,)                
+                graph = core.get_graph_dict(
+                    schema.get(key),
+                    value,
+                    subpath,
+                    options,
+                    graph)
+
+    return graph
+        
+
+def graphviz_edge(core, schema, state, path, options, graph):
+    node_spec = {
+        'name': path[-1],
+        'path': path,
+        'value': None,
+        'type': core.representation(schema)}
+
+            # if key in removed_process_keys and not retain_process_keys:
+            #     continue
+
+    graph['process_nodes'].append(node_spec)
+
+    # this is an edge, get its inputs and outputs
+    input_wires = state.get('inputs', {})
+    output_wires = state.get('outputs', {})
+    input_schema = state.get('_inputs', schema.get('_inputs', {}))
+    output_schema = state.get('_outputs', schema.get('_outputs', {}))
+
+    # bridge
+    bridge_wires = state.get('bridge', {})
+    bridge_inputs = bridge_wires.get('inputs', {})
+    bridge_outputs = bridge_wires.get('outputs', {})
+
+    # get the input and output wires
+    graph = get_graph_wires(
+        input_schema, input_wires, graph,
+        schema_key='inputs', edge_path=path, bridge_wires=bridge_inputs)
+
+    graph = get_graph_wires(
+        output_schema, output_wires, graph,
+        schema_key='outputs', edge_path=path, bridge_wires=bridge_outputs)
+
+    # get the input and output bridge wires
+    if bridge_wires:
+        # check that the bridge wires connect to valid ports
+        assert set(bridge_wires.keys()).issubset({'inputs', 'outputs'})
+
+    return graph
+
+
+visualize_types = {
+    'any': {
+        '_graphviz': graphviz_any},
+    'edge': {
+        '_graphviz': graphviz_edge},
+    'step': {
+        '_inherit': ['edge']},
+    'process': {
+        '_inherit': ['edge']}}
+
+
+# TODO: we want to visualize things that are not yet complete
+
+class VisualizeTypes(TypeSystem):
+    def __init__(self):
+        super().__init__()
+
+        self.update_types(visualize_types)
+
+
+    def get_graph_dict(self, schema, state, path, options, graph=None):
+        path = path or ()
+        graph = graph or {
+            'state_nodes': [],
+            'process_nodes': [],
+            'place_edges': [],
+            'input_edges': [],
+            'output_edges': [],
+            'disconnected_input_edges': [],
+            'disconnected_output_edges': []}
+
+        graphviz_function = self.choose_method(
+            schema,
+            state,
+            'graphviz')
+
+        return graphviz_function(
+            self,
+            schema,
+            state,
+            path,
+            options,
+            graph)
+        
+
+    def generate_graphviz(self, schema, state, path, options):
+        full_schema, full_state = self.generate(schema, state)
+        return self.get_graph_dict(full_schema, full_state, path, options)
+
+
+    def plot_graph(self,
+                   graph_dict,
+                   out_dir='out',
+                   filename=None,
+                   file_format='png',
+                   print_source=False,
+                   options=None
+                   ):
+        # make a figure
+        options = options or {}
+        graph = get_graphviz_fig(
+            graph_dict,
+            **options)
+
+        # display or save results
+        if print_source:
+            print(graph.source)
+
+        if filename is not None:
+            os.makedirs(out_dir, exist_ok=True)
+            fig_path = os.path.join(out_dir, filename)
+            print(f"Writing {fig_path}")
+            graph.render(filename=fig_path, format=file_format)
+
+
+
 # Begin Tests
 ###############
 
 plot_settings = {'out_dir': 'out',
                  'dpi': '150',}
+
+def test_graphviz():
+    cell = {
+        'config': {
+            '_type': 'map[float]',
+            'a': 11.0,  #{'_type': 'float', '_value': 11.0},
+            'b': 3333.33},
+        'cell': {
+            '_type': 'process',  # TODO -- this should also accept process, step, but how in bigraph-schema?
+            # 'config': {},
+            # 'address': 'local:cell',   # TODO -- this is where the ports/inputs/outputs come from
+            'internal': 1.0,
+            '_inputs': {
+                'nutrients': 'float',
+            },
+            '_outputs': {
+                'secretions': 'float',
+                'biomass': 'float',
+            },
+            'inputs': {
+                'nutrients': ['down', 'nutrients_store'],
+            },
+            'outputs': {
+                # 'secretions': ['secretions_store'],
+                'biomass': ['biomass_store'],
+            }
+        }
+    }
+
+    core = VisualizeTypes()
+    graphviz = core.generate_graphviz(
+        {},
+        cell,
+        (),
+        options={'dpi': '150'})
+
+    core.plot_graph(
+        graphviz,
+        out_dir='out',
+        filename='test_graphviz'
+    )
 
 
 def test_diagram_plot():
@@ -545,14 +678,22 @@ def test_diagram_plot():
             }
         }
     }
+
     plot_bigraph(cell, filename='bigraph_cell',
                  show_values=True,
                  show_types=True,
                  **plot_settings
+                 # port_labels=False,
+                 # rankdir='BT',
+                 # remove_nodes=[
+                 #     ('cell', 'address',),
+                 #     ('cell', 'config'),
+                 #     ('cell', 'interval'),
+                 # ]
                  )
 
 def test_bio_schema():
-    core = generate_types()
+    core = VisualizeTypes()
     b = {
         'environment': {
             'cells': {
@@ -694,7 +835,7 @@ def test_multi_input_output():
 
 
 def test_cell_hierarchy():
-    core = generate_types()
+    core = VisualizeTypes()
 
     core.register('concentrations', 'float')
     core.register('sequences', 'float')
@@ -756,7 +897,7 @@ def test_cell_hierarchy():
 
 
 def test_multiple_disconnected_ports():
-    core = generate_types()
+    core = VisualizeTypes()
 
     spec = {
         'process': {
@@ -780,7 +921,7 @@ def test_multiple_disconnected_ports():
         **plot_settings)
 
 def test_composite_process():
-    core = generate_types()
+    core = VisualizeTypes()
 
     spec = {
         'composite': {
@@ -830,6 +971,7 @@ def test_composite_process():
 
 
 if __name__ == '__main__':
+    test_graphviz()
     test_diagram_plot()
     test_bio_schema()
     test_flat_composite()
